Finish Responsive Navbar (#12)
This is a follow up to PR cse110-sp25-group25/restaurant-crisis#11

I made the navbar responsive on mobile using js, finishing up my work on it. I used a placeholder app icon for now, this should be replaced once the design team has the app icon ready.

I'm now done with my portion of the work on issue #9.